Use botocore instead of boto3
boto which is the old version of boto3 is already in the project
requirements, it's used by ec2-api and some other projects, and
they do not have plan to migrate boto to boto3 in pike. Add boto3
to requirements will make requirements have old and new library
at the same time. So use botocore instead of boto3 in S3 driver.

Change-Id: I777b2631b0014ba44e8fd2c409a168fd9a6ef3d2