Use allowlist_externals
This option whitelist_externals was deprecated in favor of the new
allowlist_externals option and is no longer available in recent
versions of tox.

replace() argument 1 must be str, not _StrPath

the type of app.builder.srcdir is now a _StrPath which
breaks the invocation of replace when generating
redirects.

Change-Id: I1663478fa4cf11f052bd6e772171aab885eec3d5